[#8] :monocle_face: chore(DDL): id타입  int-> bigint

+++ server/src/db/schema.sql
@@ -2,7 +2,7 @@ CREATE DATABASE IF NOT EXISTS podolog;
 USE podolog;
 
 CREATE TABLE IF NOT EXISTS `user` (
-	`userId`	int	NOT NULL auto_increment,
+	`userId`	bigint	NOT NULL auto_increment,
 	`email`	varchar(30)	UNIQUE NOT NULL,
 	`password`	varchar(60)	NULL,
 	`nickname`	varchar(40)	NOT NULL,
@@ -16,23 +16,23 @@ CREATE TABLE IF NOT EXISTS `user` (
 );
 
 CREATE TABLE IF NOT EXISTS `grainId` (
-	`grainId`	int	NOT NULL auto_increment,
-    `userId`	int	NOT NULL,
+	`grainId`	bigint	NOT NULL auto_increment,
+    `userId`	bigint	NOT NULL,
 	`grain`	int	NOT NULL DEFAULT 0,
     PRIMARY KEY (`grainId`, `userId`),
      FOREIGN KEY (`userId`) REFERENCES `user` (`userId`) ON DELETE CASCADE ON UPDATE CASCADE
 );
 
 CREATE TABLE IF NOT EXISTS `package` (
-	`packageId`	int	NOT NULL auto_increment,
+	`packageId`	bigint	NOT NULL auto_increment,
 	`packageName`	varchar(40)	NOT NULL,
      PRIMARY KEY (`packageId`)
 );
 
 CREATE TABLE IF NOT EXISTS `user_package` (
-	`userPackageId`	int	NOT NULL auto_increment,
-	`userId`	int	NOT NULL,
-	`packageId`	int	NOT NULL,
+	`userPackageId`	bigint	NOT NULL auto_increment,
+	`userId`	bigint	NOT NULL,
+	`packageId`	bigint	NOT NULL,
     `createdAt` DATETIME NOT NULL DEFAULT CURRENT_TIMESTAMP, 
     `updatedAt` DATETIME, 
     `deletedAt` DATETIME, 
@@ -42,8 +42,8 @@ CREATE TABLE IF NOT EXISTS `user_package` (
 );
 
 CREATE TABLE IF NOT EXISTS `sticker` (
-	`stickerId`	int	NOT NULL auto_increment,
-	`packageId`	int	NOT NULL,
+	`stickerId`	bigint	NOT NULL auto_increment,
+	`packageId`	bigint	NOT NULL,
 	`stickerName`	varchar(40)	NULL,
 	`stickerImg`	varchar(110) NULL,
     PRIMARY KEY (`stickerId`,`packageId`),
@@ -51,7 +51,7 @@ CREATE TABLE IF NOT EXISTS `sticker` (
 );
 
 CREATE TABLE IF NOT EXISTS `book` (
-	`bookId`	int	NOT NULL auto_increment,
+	`bookId`	bigint	NOT NULL auto_increment,
 	`bookName`	varchar(50)	NULL,
 	`color`	varchar(10)	NULL,
     `createdAt` DATETIME NOT NULL DEFAULT CURRENT_TIMESTAMP, 
@@ -61,18 +61,18 @@ CREATE TABLE IF NOT EXISTS `book` (
 );
 
 CREATE TABLE IF NOT EXISTS `user_book` (
-	`userBookId`	int	NOT NULL auto_increment,
-	`bookId`	int	NOT NULL,
-	`userId`	int	NOT NULL,
+	`userBookId`	bigint	NOT NULL auto_increment,
+	`bookId`	bigint	NOT NULL,
+	`userId`	bigint	NOT NULL,
     PRIMARY KEY (`userBookId`,`bookId`,`userId`),
     FOREIGN KEY (`bookId`) REFERENCES `book` (`bookId`) ON DELETE CASCADE ON UPDATE CASCADE,
     FOREIGN KEY (`userId`) REFERENCES `user` (`userId`) ON DELETE CASCADE ON UPDATE CASCADE
 );
 
 CREATE TABLE IF NOT EXISTS `diary` (
-	`diaryId`	int	NOT NULL auto_increment,
-	`bookId`	int	NOT NULL,
-	`userId`	int	NOT NULL,
+	`diaryId`	bigint	NOT NULL auto_increment,
+	`bookId`	bigint	NOT NULL,
+	`userId`	bigint	NOT NULL,
 	`picture`	varchar(110) NULL,
 	`title`	varchar(50)	NULL,
 	`content`	varchar(600) NULL,
@@ -85,9 +85,9 @@ CREATE TABLE IF NOT EXISTS `diary` (
 );
 
 CREATE TABLE IF NOT EXISTS `comment` (
-	`commentId`	int	NOT NULL auto_increment,
-	`userId`	int	NOT NULL,
-	`diaryId`	int	NOT NULL,
+	`commentId`	bigint	NOT NULL auto_increment,
+	`userId`	bigint	NOT NULL,
+	`diaryId`	bigint	NOT NULL,
 	`parentCommentId`	int	NOT NULL DEFAULT 0,
 	`reply`	varchar(200) NULL,
     `createdAt` DATETIME NOT NULL DEFAULT CURRENT_TIMESTAMP, 
@@ -99,10 +99,10 @@ CREATE TABLE IF NOT EXISTS `comment` (
 );
 
 CREATE TABLE IF NOT EXISTS `sticked_sticker` (
-	`stickedStickerId`	int	NOT NULL auto_increment,
-	`stickerId`	int	NOT NULL,
-	`diaryId`	int	NOT NULL,
-	`userId`	int	NOT NULL,
+	`stickedStickerId`	bigint	NOT NULL auto_increment,
+	`stickerId`	bigint	NOT NULL,
+	`diaryId`	bigint	NOT NULL,
+	`userId`	bigint	NOT NULL,
 	`positionX`	int	NULL,
 	`positionY`	int	NULL,
     `createdAt` DATETIME NOT NULL DEFAULT CURRENT_TIMESTAMP,
